Change ageInMonths again from double to long

+++ app/src/main/java/org/dhis2/usescases/enrollment/EnrollmentActivity.kt
@@ -442,7 +442,7 @@ class EnrollmentActivity : ActivityGlobalAbstract(), EnrollmentView {
         dialog.show()
     }
 
-    override fun registerBiometrics(orgUnit: String, ageInMonths: Double) {
+    override fun registerBiometrics(orgUnit: String, ageInMonths: Long) {
         BiometricsClientFactory.get(this).register(this, orgUnit, ageInMonths)
     }
